requirements.txt
opencv-python==4.10.0.84
numpy<2
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
werkzeug==3.0.1
gunicorn==21.2.0
